Move test_calculate_nondomination_rank() as non dominated sort logic moves
